feat: Add forcePopulateExercises method to database
This commit introduces a new `forcePopulateExercises` method to the `PPLWorkoutDatabase`.

This method allows for the forceful repopulation of exercises in the database. It first deletes all existing exercises and then inserts the predefined list of PPL exercises.

Logging has been added to indicate the start and success/failure of the population process.